pyproject.toml
[project]
name = "agentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.3",
    "langchain-google-genai>=4.0.0",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.4",
    "numexpr>=2.14.1",
    "python-dotenv>=1.2.1",
]
